cas/requirements.txt
pytz==2014.10
html2text==2015.2.18
pexpect==3.3
Django==1.7.5
django-braces==1.4.0
django-extensions==1.5.0
django-pagination==1.0.7
django-tastypie==0.12.1
DjangoCaptcha==0.2.13
oauthlib==0.7.2
Pillow==2.7.0
python-dateutil==2.4.1
python-mimeparse==0.1.4
six==1.9.0
django-grappelli==2.6.3